Optimize anime recommend tab
